Saving the pages with the changes made
A button is added at the bottom of the page to save the changes made by
the user on the canvas. Onclick the button, javascript fetches the
innerhtml from the mentioned id and sends it to savesvg.php. Savesvg.php
servers the function to save the content recieved into a file. In our
case, the file is called page.html.